Refactored wmap highlights into instances
